pyproject.toml
[tool.poetry]
name = "miramail"
version = "0.1.0"
description = "Email client for interacting with Mirascope"
authors = [
    "William Bakst <william@mirascope.io>",
    "Brendan Kao <brendan@mirascope.io>",
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
mirascope = "^0.7.4"
google-api-python-client = "^2.124.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
pydantic = "^2.6.4"
simplegmail = "^4.1.1"
types-python-dateutil = "^2.9.0.20240316"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
mypy = "^1.9.0"
pytest = "^8.1.1"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 88
target-version = "py38"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
exclude = ["venv", "virtualenvs"]

[[tool.mypy.overrides]]
module = "googleapiclient.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.oauth2.credentials"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.auth.transport.requests"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google_auth_oauthlib.flow"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "simplegmail"
ignore_missing_imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
